Redirect user back to EMREX Client
Fixed how the response to the EMREX Client is sent.
Now the response data is not just posted back to the client but the user is also redirected back to EMREX Client in the same instance. This is achieved by adding a html form containing the post data that is automatically submitted.

Also fixed issues with the xml. Fixed wrong value for schema location and missing leading 0 for the day for the dateTime values.